Refactor modal styles and update localization strings; remove unused admin CSS
