+++ admin/src/styles/logs.scss
@@ -242,7 +242,7 @@ button .components-spinner {
 	z-index: $modal-z-index;
 }
 
-.swpl-modal-content {
+.swpl__modal__content {
 	background-color: white;
 	border-radius: 4px;
 	box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
