--- assets/css/admin.css
@@ -1,159 +1,0 @@
-/**
- * Admin CSS
- */
-
-/* Error/Message/Note */
-.swpl-wrap div._note,
-.swpl-wrap div._error,
-.swpl-wrap div._updated,
-.swpl-wrap div._ok {
-	padding: 1px 12px;
-	margin: 5px 0 15px;
-	background-color: #fff;
-	border-left: 4px solid #d1e5ee;
-	box-shadow: 0 1px 1px 0 rgba(0, 0, 0, 0.1);
-}
-
-.swpl-wrap div._note p,
-.swpl-wrap div._error p,
-.swpl-wrap div._updated p,
-.swpl-wrap div._ok p {
-	margin: .5em 0;
-}
-
-.swpl-wrap div._error {
-	border-color: #dd3d36;
-}
-
-.swpl-wrap div._updated,
-.swpl-wrap div._ok {
-	border-color: #7ad03a;
-}
-
-.notes div._error,
-.notes div._updated,
-.notes div._ok {
-	margin: 0;
-}
-
-.notes div._note+div._note {
-	margin-top: 5px;
-}
-
-.swpl__tab__wrapper+div._note,
-.swpl__tab__wrapper+div._error,
-.swpl__tab__wrapper+div._updated,
-.swpl__tab__wrapper+div._ok {
-	margin-top: 20px;
-}
-
-/* Modal */
-#swpl-modal {
-	top: 0;
-	right: 0;
-	bottom: 0;
-	left: 0;
-	opacity: 0;
-	position: fixed;
-	overflow: hidden;
-	overflow-y: auto;
-	align-items: center;
-	z-index: 9999999999;
-	visibility: hidden;
-	transition: all .2s ease;
-	background-color: rgba(0, 0, 0, 0.5);
-}
-
-html.swpl-modal-active #swpl-modal,
-body.swpl-modal-active #swpl-modal {
-	opacity: 1;
-	visibility: visible;
-}
-
-html.swpl-modal-active>body,
-body.swpl-modal-active {
-	overflow: hidden;
-}
-
-#swpl-modal-inner {
-	width: 800px;
-	overflow: hidden;
-	max-width: 95%;
-	border-radius: 5px;
-	position: relative;
-	margin: 110px auto 40px;
-	background-color: #FFF;
-	transition: all .2s ease;
-	box-shadow: 0 10px 50px -10px rgba(0, 0, 0, 0.3);
-}
-
-body.swpl-modal-active #swpl-modal-inner {
-	margin-top: 100px;
-	margin-bottom: 50px;
-}
-
-#swpl-modal-body {
-	display: flex;
-	min-height: 360px;
-	flex-direction: column;
-}
-
-#swpl-modal-close-btn {
-	top: 15px;
-	right: 20px;
-	z-index: 10;
-	color: #999;
-	position: absolute;
-	text-decoration: none;
-}
-
-#swpl-modal-close-btn:hover,
-#swpl-modal-close-btn:focus {
-	color: #333;
-	outline: none;
-	box-shadow: none;
-}
-
-#swpl-modal-loading {
-	top: 0;
-	right: 0;
-	bottom: 0;
-	left: 0;
-	font-size: 22px;
-	display: flex;
-	position: absolute;
-	align-items: center;
-	justify-content: center;
-	background-color: #FFF;
-}
-
-.swpl-modal-header,
-.swpl-modal-content,
-.swpl-modal-footer {
-	padding: 15px 20px;
-}
-
-.swpl-modal-header {
-	font-size: 18px;
-	line-height: normal;
-	background-color: #f9f9f9;
-	border-bottom: 1px solid #eee;
-}
-
-.swpl-modal-content {
-	flex: 1 1 auto;
-	padding-top: 20px;
-	padding-bottom: 20px;
-}
-
-.swpl-modal-footer {
-	border-top: 1px solid #eee;
-}
-
-.swpl-modal-content pre {
-	margin: 0;
-	padding: 20px;
-	overflow: auto;
-	max-height: 300px;
-	background-color: #f7f7f7;
-}
