+++ assets/css/wp-debug-log.css
@@ -1,0 +1,110 @@
+/* Modal */
+#swpl-modal {
+	top: 0;
+	right: 0;
+	bottom: 0;
+	left: 0;
+	opacity: 0;
+	position: fixed;
+	overflow: hidden;
+	overflow-y: auto;
+	align-items: center;
+	z-index: 9999999999;
+	visibility: hidden;
+	transition: all .2s ease;
+	background-color: rgba(0, 0, 0, 0.5);
+}
+
+html.swpl-modal-active #swpl-modal,
+body.swpl-modal-active #swpl-modal {
+	opacity: 1;
+	visibility: visible;
+}
+
+html.swpl-modal-active>body,
+body.swpl-modal-active {
+	overflow: hidden;
+}
+
+#swpl-modal-inner {
+	width: 800px;
+	overflow: hidden;
+	max-width: 95%;
+	border-radius: 5px;
+	position: relative;
+	margin: 110px auto 40px;
+	background-color: #FFF;
+	transition: all .2s ease;
+	box-shadow: 0 10px 50px -10px rgba(0, 0, 0, 0.3);
+}
+
+body.swpl-modal-active #swpl-modal-inner {
+	margin-top: 100px;
+	margin-bottom: 50px;
+}
+
+#swpl-modal-body {
+	display: flex;
+	min-height: 360px;
+	flex-direction: column;
+}
+
+#swpl-modal-close-btn {
+	top: 15px;
+	right: 20px;
+	z-index: 10;
+	color: #999;
+	position: absolute;
+	text-decoration: none;
+}
+
+#swpl-modal-close-btn:hover,
+#swpl-modal-close-btn:focus {
+	color: #333;
+	outline: none;
+	box-shadow: none;
+}
+
+#swpl-modal-loading {
+	top: 0;
+	right: 0;
+	bottom: 0;
+	left: 0;
+	font-size: 22px;
+	display: flex;
+	position: absolute;
+	align-items: center;
+	justify-content: center;
+	background-color: #FFF;
+}
+
+.swpl-modal-header,
+.swpl-modal-content,
+.swpl-modal-footer {
+	padding: 15px 20px;
+}
+
+.swpl-modal-header {
+	font-size: 18px;
+	line-height: normal;
+	background-color: #f9f9f9;
+	border-bottom: 1px solid #eee;
+}
+
+.swpl-modal-content {
+	flex: 1 1 auto;
+	padding-top: 20px;
+	padding-bottom: 20px;
+}
+
+.swpl-modal-footer {
+	border-top: 1px solid #eee;
+}
+
+.swpl-modal-content pre {
+	margin: 0;
+	padding: 20px;
+	overflow: auto;
+	max-height: 300px;
+	background-color: #f7f7f7;
+}
